+++ NEWS.md
@@ -1,3 +1,24 @@
+# worldfootballR 0.4.8.1000 
+
+### New functions
+
+* `tm_league_debutants()` to be able to extract league debutants
+* `tm_expiring_contracts()` to be able to extract expiring contracts for a selected league
+* `tm_league_injuries()` to get all current injuries from a selected league
+* `tm_player_injury_history()` to get the full player injury history on transfermarkt for a selected player(s)
+
+
+### Improvements
+
+* `fb_player_scouting_report()` had an issue with player listed as three positions [#80](https://github.com/JaseZiv/worldfootballR/issues/80)
+* `get_match_lineups()` now also returns match and player URLs [#78](https://github.com/JaseZiv/worldfootballR/issues/78)
+* `get_match_results()` now includes match URLs [#78](https://github.com/JaseZiv/worldfootballR/issues/78)
+* `get_match_report()` now includes yellow and red card counts [#2](https://github.com/JaseZiv/worldfootballR/issues/2)
+
+### Bugs
+
+* `tm_team_player_urls()` was returning additional (incorrect) transfermarkt player URLs
+
 # worldfootballR 0.4.8
 
 ### CRAN Release